requirements.txt
pillow==10.4.0
pkgconfig==1.5.5
PyYAML==6.0.2
